requirements.txt
# Core dependencies for AutoCoder4_CC
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
fastapi>=0.100.0
httpx>=0.24.0
asyncio-extras>=1.3.0
anyio>=3.7.0

# Web server and HTTP
uvicorn[standard]>=0.22.0
aiohttp>=3.8.0
websockets>=11.0.0

# Data validation and processing
jsonschema>=4.17.0
bleach>=6.0.0
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0

# Database and async support
asyncpg>=0.28.0
databases>=0.7.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0
motor>=3.0.0

# Messaging and queues
aio-pika>=9.0.0
aiokafka>=0.8.0
aioredis>=2.0.0

# Authentication and security
bcrypt>=4.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Metrics and observability
prometheus-client>=0.17.0
sentry-sdk>=1.32.0

# AI/LLM integration
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
litellm>=1.0.0

# Template and markup processing
jinja2>=3.1.0
bleach>=6.0.0
tinycss2>=1.2.0

# Development and deployment
docker>=6.0.0
click>=8.1.0
colorama>=0.4.6
watchdog>=3.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
requests-mock>=1.10.0
hypothesis>=6.82.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities found in codebase
attrs>=23.0.0
charset-normalizer>=3.2.0
idna>=3.4
urllib3>=2.0.0
certifi>=2023.7.22